Updated openstack/openstack Project: openstack/ceilometer  f09dd12c0ef391a8f6220214109defcac78259b4
transform samples only when transformers exist

the pipeline is oddly setup to loop through each sample and
attempt to transform it even when there are no transformers. this
is madness and should be stopped -- this patch does this.

at best, this should improve performance because of needless
looping, at worst, we get rid of misleading 'tranforming sample'
logs.

Change-Id: I80ba2b7775fda03670c7fa556ad5a779c6eefd27