Updated openstack/openstack Project: openstack/ceilometer  30824df1a40346bcab8e08f5647c1c2427a9f3f2
Adapting pylint runner to the new message format

Parseable output format and --include-ids parameter are deprecated
and ignored in the pylint version used on the gate.
These functions are replaced by the --msg-template parameter.

This patch changes the runner to use the new parameter so that the
job can run properly on the gate

Closes-Bug: #1394606
Change-Id: Ia3e531923caa3f4e9a0dd03e0a4f5ef1a6d3dafb