Updated openstack/openstack Project: openstack/python-glanceclient  e386e441af9e2d3aac54f88c64bf4b93d481ba7f
Stub authentication requests rather than plugins

Rather than testing that a certain plugin was loaded test to make sure
that the client actually did the authentication request and that the
content of that request was correct.

This means we are not so tightly bound to those specific plugins and
this can be changed in later reviews.

Change-Id: Ia3ae0069e3e0277d655f0e251196eca658f94c75