Updated openstack/openstack Project: openstack/nova  e2c5aef5301688e6f135955e17ac65fdcec34022
Doc:  Adds python-tox to Ubuntu dependencies

The development.environment instructions were almost completely copy/pastable
to get started on a new machine. Almost. This doc patch just adds python-tox to
the list of things to be fetched with apt-get, so now it's really *really* simple to get
started on a brand-new machine or cloud instance.

Change-Id: I09cbb376df2ce81e618c99e5db36ddf7809f6f3d