Updated openstack/openstack Project: openstack/cinder  4750645897c30fe05fa3621a3f79850ce67e5967
Add ability to create volume from image by image name.

Some cli commands allow specifying an image by either
its name or its ID. Creating a volume from an image
currently requires knowing the image ID. The request
was made to add the ability to create a volume using
the image name instead.

This patch adds the ability to accept either name or ID.
It will attempt to locate the image based on this input
and fail if neither can be matched. If found by name, it
will also verify there are not multiple images of the
same name. If this is the case it will require using the
ID to be explicit.

Will update python-cinderclient in a separate patch once
this has been merged to allow ID or name.

Change-Id: I8b522ccff3ead644b738499e67fa85d4ab92af36
Partial-Bug: #1377823