Updated openstack/openstack Project: openstack/tempest  5e27849b3d239a93588590d10fc579dfe18fdfda
rm pyc files before test runs

It's extremely easy for incorrect local results to happen if a file
has been moved or deleted in an upstream pull. Add the rm pyc cantrip
to ensure local testing matches what a clean upstream is going to be
like as much as possible.

Change-Id: Iecf18d7c2310918ec90c14c64cd3083e46459980