Updated openstack/openstack Project: openstack/neutron  9703c40c09790480b17834fd87470c880c450144
Drop RpcProxy usage from brocade plugin

This patch drops the usage of the RpcProxy compatibility class from
the brocade plugin.  The equivalent usage of oslo.messaging APIs is
now used instead.

Note that there is one very minor functional change included.  The
previous code set the base version to '1.1'.  It should be '1.0'.  The
proper pattern to use is to set the base version on the client side to
be the initial version of the API (X.0).  Then, any methods that
require a newer version need to specify it.

Part of blueprint drop-rpc-compat.

Change-Id: I6cc343c1e88d666dd90842e5da36787b6693fc98