Updated openstack/openstack Project: openstack/cinder  200a505691af3f7bcbb02e0bfdd3ebf51f86f3c8
Add limited retype support for rbd

This patch enables limited retype support for rbd. In addition to
basic retyping between volume types that only differ in their name,
retyping between volume types that imply a change of the qos_specs
are supported. This hence allows to adapt the quality-of-service
settings of a volume after its creation.

All other changes, such as host migration, change of encryption or
changes of settings as defined by the extra_specs are not supported
by this commit and left for later patches.

Implements: blueprint ceph-rbd-support-retype
Change-Id: Ib9edf83eb3ade1e6b2dcf1121a16a6f2e68753e1