Updated openstack/openstack Project: openstack/horizon  7896990631c7016db56b1d77c9a97a4ffc263866
Support nova service-disable/enable for Hypervisor

To prevent further scheduling of instance launch on a compute we
need to run:

* nova service-disable --reason REASON NODENAME nova-compute

It would be helpful if we can send this through the hypervisors menu.
New Actions: 'Disable Service' and 'Enable Service' will be added to
Admin -> Hypervisors -> Compute Host action menu to disable/enable
service. So this is a small but useful addition to horizon.

Implement blueprint: service-disable-enable

Change-Id: Ic58d6a6410ec640aecd808ef3e586026e562ac6a