Updated openstack/openstack Project: openstack/oslo-incubator  e67f5cd0922df5a6f6e60eba7cbe3ffb849d0d77
Don't log missing policy.d as a warning

We default policy_dirs to look for a policy.d directory, but by
default that directory will not exist.  IMHO we should not be
logging a warning because of a perfectly normal default behavior.
INFO is plenty sufficient for tracking down any potential issues
with the policy dirs.

Change-Id: I183597f23b3a5eb44debd73f7a4a4444ca0d22fd