Updated openstack/openstack Project: openstack/ceilometer  8725b2ae2cab46e8c412deb12155640d2ee64c0f
Add encoding of rows and qualifiers in impl_hbase

Currently, in impl_hbase we dump only cell values.
Respectively if row or qualifier contains non-ascii symbol
exception is raised.
This patchset adds encoding/decoding for rows and qualifiers.
It uses encoding instead dumping for backward compatibility of
rows and cell keys in tables.

Change-Id: Id238d38b4ce1f91d9e5e760856e27c5f6eb0685c
Closes-bug: #1350826