Updated openstack/openstack Project: openstack/trove  2d4cd5fae1b07e09d5c70ec3736a18737795a6c5
Deletes volume on instance delete after resize

When nova deletes a instance, if the volume has been attached
during the lifetime of the instance it does not delete the volume.
If a volume is resized we detach-reattach the volume resulting in
the resource not being cleaned up after delete of trove instance.
The orphaned volumes will just pile up if there are many resize
operations.

Change-Id: I29df5b197bd90c80cf1b98361ae28d05e3c7adc2
Closes-Bug: #1326463