Updated openstack/openstack Project: openstack/oslo-incubator  f9c30017018b4deaf114f7800bc6d399ee46d16f
Add middleware.catch_errors shim for Kilo

A deprecation warning was added to oslo-incubator for Juno, but
unfortunately not to any of the projects actually using it.

Since this module is referenced in the api paste config, removing
the module from the project trees without deprecation warning
constitutes an operator-visible backwards incompatible change,
which AIUI is not allowed.

This mitigates the lack of deprecation period by adding a shim for
Kilo, which projects can sync before switching their paste defaults
to point to oslo.middleware, we can then remove the shim in L.

The module is currently used in Neutron and Swift api-paste.ini files.

Change-Id: I4ebd1baac2e3a58d086b6d227662787d42655da9
Related-Bug: #1380629