Updated openstack/openstack Project: openstack/nova  d7b6ac2938c215c25f3e6a47ede012d7511b1e78
move the integrated tests into the functional tree

This moves the integrated tests into the functional tree as an initial
seed for the functional tests. It was agreed to as an early step at
the design summit. Work on cleaning these up will happen after the
move.

PYTHON_HASH_SEED=0 is needed because these tests apparently remain
order dependent.

Also remove the use of test.ReplaceModule, which is an untested
mocking system in the nova tree that doesn't seem to reliably
work. Use stubs instead for the time being, as this will ensure
repeatability of the code.

Part of bp:functional-tests-for-nova

Change-Id: I742c487dd7bca9de3e811c545de477367fff112c