Updated openstack/openstack Project: openstack/horizon  d1c098550d0b4b631291ab93f022bd4d26ea8938
Temporarily disable the Jasmine unit tests

A recent Jasmine package upgrade to move to Jasmine 2.x is breaking the
Horizon gate. There is other work underway to make Horizon more
compatible with Jasmine 2.x (see comment #2 in the bug for more
details), let's disable the tests for now and reenable them shortly as
part of that other work when we can make sure the upgrade to 2.x
happens when the test codebase is ready to handle it.

Change-Id: I745e8957128772a7415a20393940068dbd89a493
Closes-Bug: #1396954