Updated openstack/openstack Project: openstack/oslo.messaging  0b839c57b7be932f3eb82d4916a0474013a3efbe
Notification listener pools

We can now set the pool name of a notification listener
to create multiple groups/pools of listeners consuming notifications
and that each group/pool only receives one copy of each notification.

The AMQP implementation of that is to set queue_name with the pool name.

Implements blueprint notification-listener-pools
Closes-bug: #1356226

Change-Id: I8dc0549f5550f684a261c78c58737b798fcdd656