Updated openstack/openstack Project: openstack/taskflow  f67cac77d035cc2fbadb4c85ac5b060e5c897180
Mark 'task_notifier' as renamed to 'atom_notifier'

Deprecate the usage of an engines 'task_notifier' in
favor of the more appropriately named 'atom_notifier' and
mark the 'task_notifier' property as subject to removal
in a future version.

This makes the usage of this notifier more clear since it
is not only used for task notification but also for retry
notification (which is why naming it atom notifier is more
suited to its actual usage).

Change-Id: I79d58a08fd8e6d7c8990e70bdfaae415202aa929