Updated openstack/openstack Project: openstack/oslo.db  3d5a9b134f907bc2c4ee65d6d0375927117c2e34
Repair include_object to accommodate new objects

The tests in tests/sqlalchemy/test_migrations.py which expect
to see changes in unique constraints fail due to the
assumption that the "include_object" feature of Alembic only
deals with the 'table' and 'column' constructs.   As Alembic
0.7.0 has added include_object support for 'index' and
'unique_constraint', this method should not be limiting
itself to not include objects that it's testing for.
Tests won't pass on Alembic 0.7.0 without this patch.

Change-Id: Ia71446bb8c0b248c6310534deb290524e3946987