Updated openstack/openstack Project: openstack/nova  dcb3463f234ec1a6cc69de4d1ce6ec5ddfc0e698
Check server group policy on migrate/evacuate

This patch changes server group checking in the instance group setup.
Previously, it checked for the presence of the 'group' scheduler hint
as an indication that the instance was a member of a group.  However,
that hint is only present on initial boot.  It is not present when the
scheduler is being used to pick a destination for a migration or
evactuate.

The updated logic checks the database to see if the instance is a
member of a group.  For performance reasons, it only does this is
server group functionality is enabled in the scheduler (group filters
are enabled).

This change makes it so server group policy enforcement will still
occur for migrations and evactuations, instead of only on first boot.

It also adds the call to setup_instance_group in evacuate, resize,
cold migrate and live migrate methods because now it is possible to
access the group setup.

Implements blueprint anti-affinity-on-migration

Closes-bug: #1379451

Change-Id: I6e9ea6ba69e8658fa5126a3488ee12cf2437460c