Updated openstack/openstack Project: openstack/oslo.messaging  f58d03e0ded5b251e9b6d2b3d0f5e376275df9cc
Set correctly the messaging driver to use in tests

Depending of the order of the tests executions, the driver can be the
previous configured one. But some tests are written to work on a
specific driver

This change sets the fake driver by default.

It also removes useless mock in test_log_handler.

Change-Id: Ic49bb9aec2e537e55a98315eba8dc97fbb18f96b