Updated openstack/openstack Project: openstack/neutron  8993e2654a905f2a828236dad5027971078ff58e
Remove unused variables from get_devices_details_list

Initialization of res with empty list is not needed because
res is always assigned with a new value in try/except block.

A method invocation context is created in get_device_details
and need not be initialized in get_devices_details_list because
it is never used there.

Change-Id: I2ff73390ca173cee6d73aaafdcc0c525670e358b