Updated openstack/openstack Project: openstack/neutron  580ff42286509a7e4e6ff1f07eaa49c410c1f349
Tighten up try/except block around rpc call

The try is too broad.  This commit moves code outside of the try block
where exceptions are not expected to be raised.  This only leaves the
rpc call in the current code.  In legacy code, there used to be more
but now this code merely queues updates for workers to handle
in other threads.

Change-Id: I15407553fc6ad6eeb5b483499fe1a5adffe1edca