Updated openstack/openstack Project: openstack/nova  d437f38d2385be72e507ee41e0db522683c25fb2
Fix use of conf_fixture

As the name implies, conf_fixture.ConfFixture is a fixture. It should
be used by calling self.useFixture() instead of calling it directly.
This ensures that the setUp() method is called correctly.

If the test classes are executed individually, the database config
options are not registered correctly and could result in this failure
for each test case:

NoSuchOptError: no such option in group database: backend

Closes-bug: 1396703

Change-Id: Iaf7e1aedaacf24508e55f4d6975946f83ef833a6