Updated openstack/openstack Project: openstack/sahara  68095008fa0f31a7c91c8fa12cce2cf3e97863a5
Sample JSON files for Sahara EDP APIs

Provides example payloads for the v1.1 Sahara EDP APIs, to ease
rampup for command-line users. Examples are based purely on
pre-existing EDP examples.

Documentation attempts to be as procedural as possible, while shying
away from documenting other systems. Note that though many of the flows
provided have both preconditions and step sequences in common, I
consciously chose to write each as a concretion, as any one user
is likely to focus on one example, and would likely appreciate the
all-in-one approach. There are definitely maintainability arguments
to be made against this approach, but as the 1.1 API is shipped and
stable, the extra one-time work seemed worth it.

Implementation deviates from blueprint only in that Python format
strings were used to express generated ids, as dictionary-style
formatting seemed more intuitively useful than placing uuid.EMPTY
in all cases (which would lack almost any semantics.)

As a side note, I do have a script to run these examples, but the
additional testing and maintenance overhead of adding this to the
repo seemed (to me) like overkill for this task, and would've
deviated significantly from the blueprint.

Change-Id: Ib8ab366ffc111abda4e2e064fed32d51d4a30e24
Implements: blueprint edp-api-json-samples