Updated openstack/openstack Project: openstack/horizon  eda1d97664f3c6a7079f21eb3ced0f614ef26dd0
KeyError on getting PanelGroup

When trying to get a PanelGroup at runtime and
you specify a nonregistered panel, a KeyError is
thrown when calling Dashboard.get_panel_group.
Retrieving values from a Dict using the [] formulation
results in a KeyError for non existent keys.  Using
Dict#get returns None for non existent keys which is
what this change does.

This is relevant to the dynamic panel efforts going on.

Closes-Bug: 1396682
Change-Id: I532c3be04fdd229dfd8e01b68320b2eecb6fd866