Updated openstack/openstack Project: openstack/nova  0440cd466358a5206e3aa59b19ab79532cd7f1ed
VMware: enable backward compatibility with existing clusters

When VC 5.5 creates a VM the hardware version of the VM would be set as 10.
This breaks compatibility of VM's running on an older cluster. The patch
enables the VM hardware version to be set. By default the value will not be
set. That is, it will use the current supported version. If the admin wishes
a cluster to be comptable with a specific version then they should set the
value, for example with vmx-08 to be compatible with VM's created by 5.0
and 5.1.

In the event that the cluster has mixed host version types then the VC will know
which host to place them on if the hardware version is set.

The following link describes the versions - http://goo.gl/7lhmEm

DocImpact:
The hardware version will be passed as part of the flavor extra_specs. This will
be with a key 'vmware:hw_version'

Change-Id: I67f65feceab1eab6ec4e8f4934f1c8115ed1e812
Closes-bug: #1367151