Updated openstack/openstack Project: openstack/nova  b3f8d651c21f128f7e058ef742461b8f8cfda7dd
VMWare: get storage policy from flavor

This is part of the blueprint vmware-spbm-support. The storage
policy will be passed as part of the flavor extra_specs. This will
be with a key 'vmware:storage_policy'. If there is no storage policy
on the flavor then the default will be used.

Co-authored-by: Radoslav Gerganov <rgerganov@vmware.com>

Change-Id: I75151f092e9424eaf346799072c8cf18a2784c16