Updated openstack/openstack Project: openstack/python-keystoneclient  4a49d970b06068b1ef26e7cc1f7dab3984da2ed4
Rename the client API docs

Since developers want to use the APIs, the docs should be more
enticing and say that it describes how to use the APIs.

Also, called it "V3 Client API" since this reads better than
"Client V3 API" and it matches the order in the module path.

Change-Id: I79dd6f6891bf48b477b35157256a0219426d171c