Updated openstack/openstack Project: openstack/nova  0cadd5322ca1652bae66c9e3cd94355e65396272
Removing the headroom calculation from db layer

It's odd that this is happening in the DB API because it's dealing
with instance quotas but maybe I'm not doing anything with instance
quotas, maybe I'm doing things with security group or fixed IP quotas.

This patch moves the headroom calculation from db layer to the
calling side.

Change-Id: I8379471eea4ea7b114c2a235224f6fe61f8f981f
Closes-bug: #1369696