Updated openstack/openstack Project: openstack/cinder  d496b7879517c1891a8ea260cb205c4a4694e247
Fix rpc initialization of cinder-manager volume

VolumeCommands.rpc_client is a property, so when the oslo.config argparser
introspect the class VolumeCommands at module loading time, it launch the
method. But the method depends on a initialized oslo.config.cfg.CONF object,
but this one is not yet initialized.

So don't use python property, to initialize the rpc_client correctly.

No test because of bug: #1398401

Change-Id: I2c2f0be6e7a9d0866f063d98e1f8213f62fb9f92
Closes-bug: #1398319