Updated openstack/openstack Project: openstack/nova  0a866d0e93268a7575e8fdd811fd432125b09ca6
VMware: enable VMware driver to use new BDM format

This patch makes the necessary changes in the driver to enable
it to use some of the features of the new block device mapping format.

After this patch it will be possible to set the bus, and device_type per
block device.

Change-Id: Idcf3dd97c3baacd77d03616b5b2dd768acf46de4