Updated openstack/openstack Project: openstack/nova  f568b267f7e9b0d9b89c73c352a1ba5ddf64563d
Document the upgrade plans

This covers the plan for zero downtime upgrades. It looks at all the
building blocks we have in place, and a few we have planned for kilo and
pulls that into a process that we hope will deliver zero downtime
upgrades.

This is just an initial stake in the ground to help drive the upgrade
efforts during kilo.

Change-Id: I3974d18f79ba655257dcac49208821bc191d8cc5