Updated openstack/openstack Project: openstack/nova  e430443f8b1394c5a3ca715f6e9437c6c00bcf36
add abandon_old_reviews script

This is an automated script to abandon the reviews that we agreed as a
community are beyond stale.

No activity for 4w with a -2 on them or No activity for 4w and Jenkins
doesn't have a +1 or better on it.

Can be run by any core reviewer.

Change-Id: I19636f04b691c0b74af306709d54b26978c19282