Updated openstack/openstack Project: openstack/nova  13ed48d1968bae11a977e0cf10cda3f84ffe31a9
remove test.ReplaceModule from test.py

The ReplaceModule fixture is buggy and very sensitive to the sequence
things are imported in. It was only used in 3 places until the
functional test split landed (and had to be removed from there).

This removes the use of ReplaceModule for fake_ldap. Unwinding the
xenapi test code is a bit more challenging especially given how
spidery all that code is. So instead of fixing that move the
ReplaceModule into the xenapi test code where it's used and leave it
as an exercise for the xenapi folks to move to more real mocks or
monkey patching.

This gets it out of test.py so that no one else will accidentally use
it.

Change-Id: I2398a83039ef50a926c1adecbb680c935bae7ed8