Updated openstack/openstack Project: openstack/nova  9537e3cee2321a54e7aa736497003219b2e349aa
extract _setup_timeouts in test.py

Cleaning up the base test.py to be a little more understandable about
what is going on, the chunks of work being accomplished, and why. This
will hopefully make it possible to pull some of this work out of tests
that don't need it.

Change-Id: I2baa75280c32e5a2affcdd54d231df7f034d390f